added constants, added verification method to OTP for reset password, updated request body for registering user, added EmailSendingException,  added OTPGenerator, added authority check to methods, added userDetailsService implementation,

+++ src/main/java/com/presidio/rentify/controller/PropertyController.java
@@ -2,14 +2,15 @@ package com.presidio.rentify.controller;
 
 
 import com.presidio.rentify.constants.AppConstants;
+import com.presidio.rentify.dto.APIResponse;
 import com.presidio.rentify.dto.PageableResponse;
 import com.presidio.rentify.dto.PropertyDTO.PropertyRequestDTO;
 import com.presidio.rentify.dto.PropertyDTO.PropertyResponseDTO;
+import com.presidio.rentify.security.AuthenticationFacade;
 import com.presidio.rentify.service.PropertyService;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
@@ -23,51 +24,86 @@ public class PropertyController {
     @Autowired
     private PropertyService propertyService;
 
-    @PreAuthorize("hasRole('SELLER')")
+    @Autowired
+    private AuthenticationFacade authenticationFacade;
+
+    @PreAuthorize("hasAuthority('SELLER')")
     @PostMapping("/user/{userId}")
-    public ResponseEntity<PropertyResponseDTO> addProperty(@PathVariable Long userId, @Valid @RequestBody PropertyRequestDTO propertyRequestDTO) {
-        PropertyResponseDTO property = propertyService.addProperty(userId, propertyRequestDTO);
-        return ResponseEntity.ok(property);
+    public ResponseEntity<APIResponse<PropertyResponseDTO>> addProperty(@PathVariable Long userId, @Valid @RequestBody PropertyRequestDTO propertyRequestDTO) {
+
+        String authenticatedUsername = authenticationFacade.getAuthenticatedUsername();
+        if (propertyService.isOwner(userId, authenticatedUsername)) {
+            PropertyResponseDTO property = propertyService.addProperty(userId, propertyRequestDTO);
+            return ResponseEntity.ok(new APIResponse<>(true, "Property added successfully", property));
+        }
+        return new ResponseEntity<>(new APIResponse<>(false, "Access denied"), HttpStatus.FORBIDDEN);
+
     }
 
-    @PreAuthorize("hasRole('SELLER')")
+    @PreAuthorize("hasAuthority('SELLER')")
     @GetMapping("/owner/{ownerId}")
-    public ResponseEntity<PageableResponse<PropertyResponseDTO>> getPropertiesByOwner(
+    public ResponseEntity<APIResponse<PageableResponse<PropertyResponseDTO>>> getPropertiesByOwner(
                                                                         @RequestParam(value = "pageNo", defaultValue = AppConstants.PAGE_NUMBER, required = false) Integer pageNo,
                                                                         @RequestParam(value = "pageSize", defaultValue = AppConstants.PAGE_SIZE, required = false) Integer pageSize,
                                                                         @PathVariable Long ownerId) {
-        PageableResponse<PropertyResponseDTO> properties = propertyService.getPropertiesByOwner(pageNo, pageSize, ownerId);
-        return ResponseEntity.ok(properties);
+
+        String authenticatedUsername = authenticationFacade.getAuthenticatedUsername();
+        if (propertyService.isOwner(ownerId, authenticatedUsername)) {
+            PageableResponse<PropertyResponseDTO> properties = propertyService.getPropertiesByOwner(pageNo, pageSize, ownerId);
+            return ResponseEntity.ok(new APIResponse<>(true, "Properties fetched successfully", properties));
+        }
+        return new ResponseEntity<>(new APIResponse<>(false, "Access denied"), HttpStatus.FORBIDDEN);
+
+
     }
 
-    @PreAuthorize("hasRole('BUYER')")
+    @PreAuthorize("hasAuthority('BUYER')")
     @GetMapping
-    public ResponseEntity<PageableResponse<PropertyResponseDTO>> getAllProperties(
+    public ResponseEntity<APIResponse<PageableResponse<PropertyResponseDTO>>> getAllProperties(
                                                                         @RequestParam(value = "pageNo", defaultValue = AppConstants.PAGE_NUMBER, required = false) Integer pageNo,
                                                                         @RequestParam(value = "pageSize", defaultValue = AppConstants.PAGE_SIZE, required = false) Integer pageSize,
                                                                         @RequestParam Map<String, String> filters) {
+
         PageableResponse<PropertyResponseDTO> properties = propertyService.getAllProperties(pageNo, pageSize, filters);
-        return ResponseEntity.ok(properties);
+        return ResponseEntity.ok(new APIResponse<>(true, "Properties fetched successfully", properties));
+
     }
 
-    @PreAuthorize("hasRole('SELLER')")
+    @PreAuthorize("hasAuthority('SELLER')")
     @PutMapping("/{propertyId}")
-    public ResponseEntity<PropertyResponseDTO> updateProperty(@PathVariable Long propertyId, @Valid @RequestBody PropertyRequestDTO propertyRequestDTO) {
-        PropertyResponseDTO updatedProperty = propertyService.updateProperty(propertyId, propertyRequestDTO);
-        return ResponseEntity.ok(updatedProperty);
+    public ResponseEntity<APIResponse<PropertyResponseDTO>> updateProperty(@PathVariable Long propertyId, @Valid @RequestBody PropertyRequestDTO propertyRequestDTO) {
+
+        String authenticatedUsername = authenticationFacade.getAuthenticatedUsername();
+        if (propertyService.isOwnerByPropertyId(propertyId, authenticatedUsername)) {
+            PropertyResponseDTO updatedProperty = propertyService.updateProperty(propertyId, propertyRequestDTO);
+            return ResponseEntity.ok(new APIResponse<>(true, "Property updated successfully", updatedProperty));
+        }
+        return new ResponseEntity<>(new APIResponse<>(false, "Access denied"), HttpStatus.FORBIDDEN);
+
     }
 
-    @PreAuthorize("hasRole('SELLER')")
+    @PreAuthorize("hasAuthority('SELLER')")
     @DeleteMapping("/{propertyId}")
-    public ResponseEntity<Void> deleteProperty(@PathVariable Long propertyId) {
-        propertyService.deleteProperty(propertyId);
-        return ResponseEntity.noContent().build();
+    public ResponseEntity<APIResponse<Void>> deleteProperty(@PathVariable Long propertyId) {
+
+        String authenticatedUsername = authenticationFacade.getAuthenticatedUsername();
+        if (propertyService.isOwnerByPropertyId(propertyId, authenticatedUsername)) {
+            propertyService.deleteProperty(propertyId);
+            return ResponseEntity.ok(new APIResponse<>(true, "Property deleted successfully"));
+        }
+        return new ResponseEntity<>(new APIResponse<>(false, "Access denied"), HttpStatus.FORBIDDEN);
+
     }
 
-    @PreAuthorize("isAuthenticated()")
+    @PreAuthorize("hasAuthority('BUYER')")
     @PostMapping("/{propertyId}/like")
-    public ResponseEntity<PropertyResponseDTO> likeProperty(@PathVariable Long propertyId) {
-        PropertyResponseDTO likedProperty = propertyService.likeProperty(propertyId);
-        return ResponseEntity.ok(likedProperty);
+    public ResponseEntity<APIResponse<PropertyResponseDTO>> likeProperty(@PathVariable Long propertyId) {
+
+        String authenticatedUsername = authenticationFacade.getAuthenticatedUsername();
+        if (propertyService.isOwnerByPropertyId(propertyId, authenticatedUsername)) {
+            PropertyResponseDTO likedProperty = propertyService.likeProperty(propertyId);
+            return ResponseEntity.ok(new APIResponse<>(true, "Property liked successfully", likedProperty));
+        }
+        return new ResponseEntity<>(new APIResponse<>(false, "Access denied"), HttpStatus.FORBIDDEN);
     }
 }
