+++ src/main/java/com/presidio/rentify/controller/UserController.java
@@ -5,8 +5,8 @@ import com.presidio.rentify.dto.UserDTO.PasswordUpdateDTO;
 import com.presidio.rentify.dto.UserDTO.UserRequestDTO;
 import com.presidio.rentify.dto.UserDTO.UserResponseDTO;
 import com.presidio.rentify.dto.UserDTO.UserResponseWithTokenDTO;
+import com.presidio.rentify.security.AuthenticationFacade;
 import com.presidio.rentify.service.UserService;
-import com.presidio.rentify.service.impl.UserServiceImpl;
 import jakarta.validation.Valid;
 import org.modelmapper.ModelMapper;
 import org.slf4j.Logger;
@@ -14,9 +14,9 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.Map;
 import java.util.Optional;
 
 @RestController
@@ -31,28 +31,58 @@ public class UserController {
     @Autowired
     private ModelMapper modelMapper;
 
+    @Autowired
+    private AuthenticationFacade authenticationFacade;
+
+    @PreAuthorize("isAuthenticated()")
     @GetMapping("/{id}")
     public ResponseEntity<APIResponse<Optional<UserResponseDTO>>> getUserDetails(@PathVariable Long id) {
+        String username = authenticationFacade.getAuthenticatedUsername();
         Optional<UserResponseDTO> user = userService.findUserById(id);
-        return new ResponseEntity<>(new APIResponse<>(true, "User fetched successfully", user), HttpStatus.OK);
+        if (user.isPresent() && user.get().getEmail().equals(username)) {
+            return new ResponseEntity<>(new APIResponse<>(true, "User fetched successfully", user), HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(new APIResponse<>(false, "Access denied"), HttpStatus.FORBIDDEN);
+        }
     }
 
+    @PreAuthorize("isAuthenticated()")
     @PutMapping("/{id}")
     public ResponseEntity<APIResponse<UserResponseDTO>> updateUserDetails(@PathVariable Long id, @Valid @RequestBody UserRequestDTO userRequestDTO) {
-        UserResponseDTO updatedUser = userService.updateUser(id, userRequestDTO);
-        return ResponseEntity.ok(new APIResponse<>(true, "User updated successfully", updatedUser));
+        String username = authenticationFacade.getAuthenticatedUsername();
+        Optional<UserResponseDTO> user = userService.findUserById(id);
+        if (user.isPresent() && user.get().getEmail().equals(username)) {
+            UserResponseDTO updatedUser = userService.updateUser(id, userRequestDTO);
+            return ResponseEntity.ok(new APIResponse<>(true, "User updated successfully", updatedUser));
+        } else {
+            return new ResponseEntity<>(new APIResponse<>(false, "Access denied"), HttpStatus.FORBIDDEN);
+        }
     }
 
+    @PreAuthorize("isAuthenticated()")
     @PutMapping("/{id}/update-password")
     public ResponseEntity<APIResponse<UserResponseWithTokenDTO>> updatePassword(@PathVariable Long id, @Valid @RequestBody PasswordUpdateDTO passwordUpdateDTO) {
-        UserResponseWithTokenDTO userResponseDTO = userService.updatePassword(id, passwordUpdateDTO);
-        return ResponseEntity.ok(new APIResponse<>(true, "Password updated successfully", userResponseDTO));
+        String username = authenticationFacade.getAuthenticatedUsername();
+        Optional<UserResponseDTO> user = userService.findUserById(id);
+        if (user.isPresent() && user.get().getEmail().equals(username)) {
+            UserResponseWithTokenDTO userResponseDTO = userService.updatePassword(id, passwordUpdateDTO);
+            return ResponseEntity.ok(new APIResponse<>(true, "Password updated successfully", userResponseDTO));
+        } else {
+            return new ResponseEntity<>(new APIResponse<>(false, "Access denied"), HttpStatus.FORBIDDEN);
+        }
     }
 
+    @PreAuthorize("isAuthenticated()")
     @DeleteMapping("/{id}")
     public ResponseEntity<APIResponse<Void>> deleteUser(@PathVariable Long id) {
-        userService.deleteUser(id);
-        return ResponseEntity.ok(new APIResponse<>(true, "User deleted successfully"));
+        String username = authenticationFacade.getAuthenticatedUsername();
+        Optional<UserResponseDTO> user = userService.findUserById(id);
+        if (user.isPresent() && user.get().getEmail().equals(username)) {
+            userService.deleteUser(id);
+            return ResponseEntity.ok(new APIResponse<>(true, "User deleted successfully"));
+        } else {
+            return new ResponseEntity<>(new APIResponse<>(false, "Access denied"), HttpStatus.FORBIDDEN);
+        }
     }
 
 }
